* adding support for additional operations * milestone will be creating a testbed that can debugged and optimized for supported platforms * platforms are: Windows 32/64, Linux 32/64 and OS X (64-bit) - required 3 different JIT compilers

+++ src/jit/amd64/win/jit_amd_lp64.h
@@ -237,7 +237,7 @@ namespace jit {
     unsigned char* code;
     long code_index;   
     double* floats;     
-    long floats_index;
+    size_t floats_index;
     size_t instr_index;
     long code_buf_max;
     bool compile_success;
